dom/html/test/test_bug242709.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=242709
-->
<head>
  <title>Test for Bug 242709</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=242709">Mozilla Bug 242709</a>
<p id="display"></p>
<div id="content">
<iframe src="bug242709_iframe.html" id="a"></iframe> 
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 242709 **/

SimpleTest.waitForExplicitFinish();

var submitted = function() {
  ok(true, "Disabling button after form submission doesn't prevent submitting");
  SimpleTest.finish();
}

</script>
</pre>
</body>
</html>
